+++ packages/data-grid/src/cell-renderers/number-cell-renderer.models.ts
@@ -1,0 +1,4 @@
+export interface NumberCellRendererConfig {
+  /** Number of decimal places */
+  decimals?: number;
+}